requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.3.0
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
braceexpand==0.1.7
cachetools==5.5.0
cashews==7.3.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
config==0.5.1
dataclass-factory==2.16
discord-ios==1.1.1
flatbuffers==24.3.25
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
googletrans==3.0.0
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
humanfriendly==10.0
humanize==4.10.0
hyperframe==5.2.0
idna==2.10
imageio==2.35.1
import_expression==1.1.5
jishaku==2.5.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.43.0
meow.py @ git+https://github.com/parelite/discord.py@ef0d2896bc09f72a5eeaf5ee3421109044fd7b27
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numba==0.60.0
numpy==1.26.4
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
orjson==3.10.7
packaging==24.1
pillow==10.4.0
platformdirs==4.3.1
pomice==2.9.1
pooch==1.8.2
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==1.10.18
pydub==0.25.1
PyMatting==1.1.12
pyparsing==3.1.4
python-dotenv==1.0.1
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.20.0
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
shazamio==0.6.0
shazamio_core==1.0.7
six==1.16.0
sniffio==1.3.1
sympy==1.13.2
tifffile==2024.8.30
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uwuipy==0.1.9
websockets==13.0.1
yarl==1.10.0
